requirements.txt
streamlit
fastapi
groq
uvicorn
langchain
langchain-groq
python-dotenv